+++ application/modules/email/models/Emails.php
@@ -1,8 +1,0 @@
-<?php
-
-class Email_Model_Emails
-{
-
-
-}
-