+++ public/layouts/default/js/core.js
@@ -5,6 +5,32 @@
 var requireSymbol = ' *';
 
 /**
+ * Show error modal
+ * @param {object} option
+ *  heading:'Ooops',
+ *  body:'Error text'
+ * @returns {mixed}
+ */
+function errorModal(options) {
+    var context = $('#errorModal');
+
+    var defaults = {
+        heading: null,
+        body: null
+    };
+
+    var options = $.extend(defaults, options);
+    if (options.heading != null) {
+        $(context).find('.modal-header h3').text(options.heading);
+    }
+    if (options.body != null) {
+        $(context).find('.modal-body p').text(options.body);
+    }
+
+    $(context).modal('show');
+}
+
+/**
  * Show confirm modal
  * @param {object} option
  *  heading:'Confirm your action',